Refactor to modular Canvas and Dots
Modular refactor of Canvas and Dots -
Advanced in successfully passing the data information from the TextBox to the Dots class through the Canvas class.

Also created instance of P5 in Canvas that uses p to call the P5 object in functions located in Dots, and will be used in future refactoring.